requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-Login==0.6.3
PyMySQL==1.1.0
cryptography==41.0.4
bcrypt==4.0.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
onnxruntime==1.16.1
Werkzeug==2.3.7
python-dateutil==2.8.2
